Created a new Expo app
