--- .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,0 @@
-blank_issues_enabled: false
-contact_links:
-  - name: Rodrigo Cordeiro
-    url: rodrigomendoncca@gmail.com
-    about: Please ask and answer questions here.
