--- .github/ISSUE_TEMPLATE/issue_template.md
@@ -1,10 +1,0 @@
----
-name: Blank issue
-about: Could be used for general issues
-title: ''
-labels: ''
-assignees: ''
----
-
-Please add your information bellow
-
