--- .github/pull_request_template.md
@@ -1,11 +1,0 @@
-# Description
-
- What is this pull request for? Add a little description about what was done and why. Add images or gists if needed.
-
-# Type of PR
- - [ ] Bugfix
- - [ ] New feature
- - [ ] Documentation
-
-# Changes made
- What have you added/removed to the code.
