--- .github/workflows/CodeAnalysis.yml
@@ -1,44 +1,0 @@
-name: "Scan code and validate build"
-
-on:
-  push:
-    branches: [
-      "main",
-      "feat/**",
-      "bug/**",
-      "develop"
-    ]
-  pull_request:
-    # The branches below must be a subset of the branches above
-    branches: [ main ]
-  schedule:
-    - cron: '0 3 * * 1'
-
-jobs:
-  analyze:
-    name: Analyze
-    runs-on: ubuntu-latest
-    permissions:
-      actions: read
-      contents: read
-      security-events: write
-    strategy:
-      fail-fast: false
-      matrix:
-        language: [ 'javascript' ]
-        
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v2
-    
-    # Initializes the CodeQL tools for scanning.
-    - name: Initialize CodeQL
-      uses: github/codeql-action/init@v1
-      with:
-        languages: ${{ matrix.language }}
-        
-    - name: Autobuild
-      uses: github/codeql-action/autobuild@v1
-    
-    - name: Perform CodeQL Analysis
-      uses: github/codeql-action/analyze@v1
