--- .github/workflows/PublishUpdates.yml
@@ -1,38 +1,0 @@
-name: Expo actions
-on:
-  push:
-    branches:
-      - master
-jobs:
-  Publish_Updates:
-    name: Install and publish
-    if: "${{github.repository_owner == 'darthC0de'}}"
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: 14.x
-      - uses: expo/expo-github-action@v6
-        with:
-          expo-version: 4.x
-          expo-cache: true
-          token: ${{ secrets.EXPO_TOKEN }}
-      - run: npm install
-      - run: expo publish
-  GenerateNewBuild:
-    name: Build new artifact
-    runs-on: ubuntu-latest
-    needs: Publish_Updates
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: 14.x
-      - uses: expo/expo-github-action@v6
-        with:
-          expo-version: 4.x
-          expo-cache: true
-          token: ${{ secrets.EXPO_TOKEN }}
-      - run: npm install
-      - run: expo build:android -t apk --no-wait
