--- .github/workflows/ValidatesBranchNaming.yml
@@ -1,14 +1,0 @@
-name: 'Assert Branch Naming Convention'
-on: [ push, pull_request ]
-
-jobs:
-  branch-naming-rules:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: deepakputhraya/action-branch-name@master
-        with:
-          regex: '\[([\w]{3,})\]+\:+([\w]{5,})' # Regex the branch should match. This example enforces grouping
-          allowed_prefixes: 'feature,stable,fix,updates' # All branches should start with the given prefix
-          ignore: master,develop,main # Ignore exactly matching branch names from convention
-          min_length: 5 # Min length of the branch name
-          max_length: 20 # Max length of the branch name
