+++ .gitignore
@@ -1,5 +1,5 @@
-node_modules/**/*
-.expo/*
+node_modules/
+.expo/
 npm-debug.*
 *.jks
 *.p8
@@ -11,7 +11,3 @@ web-build/
 
 # macOS
 .DS_Store
-
-
-#yarn files
-yarn*
