--- README.md
@@ -1,13 +1,0 @@
-# Kiss Manager
-![Icon](assets/icon@3x.png) 
-
-![Top Language](https://img.shields.io/github/languages/top/JuniorAlvess/kissManager?style=flat-square)
-![License](https://img.shields.io/github/license/JuniorAlvess/kissManager?color=lightblue&style=flat-square)
-![Website](https://img.shields.io/website?down_color=red&down_message=Got%20Shot&style=flat-square&up_color=green&up_message=Up&url=https%3A%2F%2Fwww.rodcordeiro.com.br)
-[![Expo Publish](https://github.com/darthC0de/kissManagerMobile/actions/workflows/PublishUpdates.yml/badge.svg)](https://github.com/darthC0de/kissManagerMobile/actions/workflows/PublishUpdates.yml)
-
-![React Native](https://img.shields.io/badge/React%20Native-black?style=flat-square&logo=react)
-![typescript](https://img.shields.io/badge/Typescript-black?style=flat-square&logo=typescript)
-![Javascript](https://img.shields.io/badge/Javascript-black?style=flat-square&logo=javascript)
-
-The kiss manager is a password manager created to help users to use and manage secure passwords, being able to improve their accounts security.
