--- UserFluxogram.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2021-07-01T15:57:21.359Z" agent="5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.6.13 Chrome/89.0.4389.128 Electron/12.0.7 Safari/537.36" etag="t4IZIXJfiiz6dqCbK2vr" version="14.6.13" type="device"><diagram id="WLeNL2iAT3cLDlL8Z_Fv" name="Page-1">7Vtdc5s6EP01fkwGJIOdR8dJ2s40M+l1px+PslFADSCukBOnv/5KIL4kHJPGGHrb6YyLFiFg95zdI6FM4DLavWMoCW6ph8MJsLzdBF5NAJi7c/ErDc+5wYbQzS0+I56yVYYV+YmV0VLWLfFw2ujIKQ05SZrGDY1jvOENG2KMPjW73dOwedcE+dgwrDYoNK1ficcDZbXdi+rEe0z8QN16Dmb5iTXaPPiMbmN1v5jGOD8ToWIY9Y5pgDz6VDPB6wlcMkp5fhTtljiUbi08ll93s+ds+cgMx7zLBV+urU+ffJjMv92DB5+/t1fW1zOYj/KIwq1yhXpY/lz4BnvCVapJGQ+oT2MUXlfWy+z9sbyNJVrpA+Yb+eq2aFQXfKQ0UcYfmPNnBQK05VSYAh6F6qz5XgUiEPMxf+Fl1LPLB65dqLzxDtMIc/YsOjAcIk4em7FHCkJ+2a+89I4S8SjAUnB3XRVRhfYz58JqjpHSLdtgdVkVEnFQe47KlAXqFUGz7cGDhHeEf5NjnTuq9b125mqnbpM1nlVjb2Bzdx0O7EEAFOmkbwTMoIYAONUQkD9qbwgABm1XIu24KBLRu4zXaZK52w2Fsy6JOPB5zbAuDHiHN1suMKFOiEdZ652FrTaAhjoW0Gi9Fa98+RQQjlcJysL4JIpEEy4vgG8vKh4x43j3YhgLQlqORshijKd6MlchCmp5vDS2xb4RtNdGaNqSWHOPBrbu4zRBccOx7r9bWRUuxfvzMxQSP57AhSx/wkeYVef1ge5phlJzIHniLM3ILMex3WRnjvJF3MlDkyWcLGbZL6SypssflCQ1QOS3MXCSv4aGNdbS0fRAUhgWuc9efA6LsMxs5yAh92SDmHxtQQGr6vWj7HRPieQ89VA2SMLwI0GR9OUELKVB6JmIpuJoy0lIfmajyZ4hFRJhxSmT8iEDKcoER4R+4jjrtY1Q9sSM5Hd7xOG5MCxCn2ZyJZL/Va6uMa50vuwlR83SUprQlMtDzra45rJkLwclRppkSzmjD3hJQ8qEJVMkAgEkDDWTiNeGxL4wOFXrc1YAzoC1h9RUePw+zHRMQDwPx72zezbVy20Lu0Eru0Ff7HbM/CsyvuTeQmJPC5F8V4HR8CNa4/COpoQTKvx2taaci9jDy6LDIqf6Fada/iy1o3Swh9Kg8nYW7EK0qr6JvHG086VYP0exxyjxzpNAhB0Y+JgAuLHkvz4SM9BCN5tbRujmbZFz+4qc20Hwxt5CzilEaxOiNCWbPdqnpnwqHTSU9pl1lD41zzstni9sb1RIjk7aaTEjPKCRzZF0DE1dbaSetZbdVsrHOkc6CKdZVykNjy2l30TbmRGDGxSm+IjVUMmsK6WxzKQciXKXhbKtLDbDe4TU6Vg67GeukTqnLQTureZBs+j9xjyYd+QBnI6KB3MjBreIxHeZPtViMfCUrAlfYNb9Up3V0dtb3beHB6vwJnuuLZrI5vdCLMhGJR2y1tu1QyGKD6J8PiqQF49dQ/nnbDo0ymR/dOro0qmc2wyV+W0z6wxOpjeQAnaVQF0l9YlYYa7VLxKGU+GAFxZLXlcUjo1lMNNVzIWpYk5bCICZXAYH8wCVoes8oBix7yV1qC+pA/vE0zxzirFEHpIVRVhXOA7QLRatVwut3jkFncE55YyIQg0CVXw6OoW6f5Yc1xQCzEcUrFPlu+7BGlfRL9ND80uf9ZluN8EH0bToVvzciOTz4cqI60i+z4GZ9n2ulK/DfZ8D7ohY0ExZoMecNe1Ig3FNCIG5BDvm5b9j82eqbzixLxyDP6ddCxyBZj7tDpSuermQsb3rZf2ziC7uet6DBEy5/Aet0UBb837Ld83TEhIOT8g95WzWXzmDg61v7tkZqK16O9oIPU9hobm08Q/2iSBctlel1Iw3e+TiwCKx6TtolrjTTmqhKTqWDCOO71CaPlHmjct/c00kDP7FBZpbLW5RjPzSf+m4HKiTt9yDMpwHzXX3FXocLf7AyNw3NdOhWSTb9vqkci9Xpy1AxXGXLUCnX6EqGHhYtHatj6fZPDRzNCbOtSG67h2a65SGpy3IU3P15mgI3LOEZh9AU/u0yXoRuSdA4F8A9gNA8KcD8Oj7fP4i8HUINL9aX3uE/yYaEFotf0hyWhHTZffp/5nB067bNYpaMxYGT5sfPaADfo3BrtMc6HirCqJZ/flp3r368154/R8=</diagram></mxfile>
