+++ app.json
@@ -2,20 +2,16 @@
   "expo": {
     "name": "kissManager",
     "slug": "kissManager",
-    "version": "2.0.0",
-    "description": "Gerenciador de senhas",
-    "owner": "darthcode",
-    "githubUrl": "https://github.com/rodcordeiro/kissManagerMobile",
+    "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
-    "backgroundColor": "#00587A",
     "splash": {
       "image": "./assets/splash.png",
-      "resizeMode": "cover",
-      "backgroundColor": "#00587A"
+      "resizeMode": "contain",
+      "backgroundColor": "#ffffff"
     },
     "updates": {
-      "fallbackToCacheTimeout": 30000
+      "fallbackToCacheTimeout": 0
     },
     "assetBundlePatterns": [
       "**/*"
@@ -23,19 +19,14 @@
     "ios": {
       "supportsTablet": true
     },
+    "android": {
+      "adaptiveIcon": {
+        "foregroundImage": "./assets/adaptive-icon.png",
+        "backgroundColor": "#FFFFFF"
+      }
+    },
     "web": {
       "favicon": "./assets/favicon.png"
-    },
-    "android": {
-      "googleServicesFile": "./google-services.json",
-      "adaptiveIcon": {
-        "foregroundImage": "./assets/icon.png",
-        "backgroundColor": "#A9F448"
-      },
-      "package": "com.darthcode.kissManager"
-    },
-    "notification":{
-      "color":"#00587A"
     }
   }
 }
