+++ package.json
@@ -8,53 +8,18 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/clipboard": "1.5.1",
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-native-community/picker": "1.6.6",
-    "@react-navigation/native": "^5.7.6",
-    "@react-navigation/stack": "^5.9.3",
-    "eslint-import-resolver-node": "0.3.4",
-    "expo": "^41.0.0",
-    "expo-linear-gradient": "~8.3.0",
-    "expo-sqlite": "~8.4.0",
-    "expo-status-bar": "~1.0.2",
-    "expo-updates": "~0.2.1",
-    "react": "^17.0.1",
+    "expo": "~42.0.1",
+    "expo-status-bar": "~1.0.4",
+    "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.3.tar.gz",
-    "react-native-flash-message": "0.1.17",
-    "react-native-gesture-handler": "1.7.0",
-    "react-native-item-select": "0.3.0",
-    "react-native-picker-select": "8.0.3",
-    "react-native-reanimated": "2.1.0",
-    "react-native-screens": "2.10.1",
-    "react-native-web": "0.15.7",
-    "react-native-web-hooks": "3.0.1",
-    "react-navigation": "^4.4.2"
+    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "react-native-web": "~0.13.12"
   },
   "devDependencies": {
-    "@babel/core": "~7.9.0",
-    "@types/react": "16.9.53",
-    "@types/react-dom": "16.9.8",
-    "@types/react-native": "0.63.25",
-    "typescript": "3.9.7"
+    "@babel/core": "^7.9.0",
+    "@types/react": "~16.9.35",
+    "@types/react-native": "~0.63.2",
+    "typescript": "~4.0.0"
   },
-  "private": true,
-  "renovate": {
-    "branchPrefix": "[Updates]:",
-    "packageRules": [
-      {
-        "updateTypes": [
-          "minor",
-          "patch",
-          "pin",
-          "digest"
-        ],
-        "automerge": true
-      }
-    ],
-    "extends": [
-      "config:base"
-    ]
-  }
+  "private": true
 }
