--- src/database/connection.tsx
@@ -1,5 +1,0 @@
-import * as SQLite from 'expo-sqlite';
-
-export const DatabaseConnection = {
-    getConnection: () => SQLite.openDatabase("kmdb.db"),
-};
